+++ src/modules/src/log.c
@@ -896,7 +896,7 @@ void logRunBlock(void * arg)
       if (blk->droppedPackets++ % 100 == 0)
       {
         DEBUG_PRINT("WARNING: LOG packets drop detected (%lu packets lost)\n",
-                    blk->droppedPackets);
+                     (unsigned long)blk->droppedPackets);
       }
     }
   }
